+++ packages/cascara/src/modules/AllModules.fixture.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import FormProvider from '../ui/Form/context/FormProvider';
-import tableStyles from '../ui/Table/Table.module.scss';
+import FormProvider from '../ui/FormOld/context/FormProvider';
+import tableStyles from '../ui/TableOld/TableOld.module.scss';
 import {
   ActionButton,
   DataCheckbox,
@@ -13,7 +13,7 @@ import {
   ModuleSandbox,
 } from './';
 
-import FormActionEdit from '../ui/Form/modules/ActionEdit';
+import FormActionEdit from '../ui/FormOld/modules/ActionEdit';
 // import TableActionEdit from '../ui/Table/modules/ActionEdit';
 
 const lorem =